Create deployment.yaml
In Kubernetes (and Minikube) is primarily used to define and manage the deployment of applications. It describes how many replicas of a pod should run, which container image to use, and how updates should be handled. My version is very basic, as the task did not specified any complicated or adavnced configurations.